Ignore MultilingualDisabled test when Multilingual is enabled
